+++ shared/src/main/scala/scopt/RenderingMode.scala
@@ -1,0 +1,7 @@
+package scopt
+
+trait RenderingMode
+object RenderingMode {
+  case object OneColumn extends RenderingMode
+  case object TwoColumns extends RenderingMode
+}
